+++ cpu/isr.c
@@ -1,6 +1,7 @@
 #include "isr.h"
 #include "idt.h"
 #include "../drivers/screen.h"
+#include "../drivers/keyboard.h"
 #include "../libc/string.h"
 #include "timer.h"
 #include "ports.h"
@@ -148,4 +149,6 @@ void irq_install() {
     asm volatile("sti");
     /* IRQ0: timer */
     //init_timer(50);
+    /* IRQ1: keyboard */
+    init_keyboard();
 }
