+++ cpu/soundManager.h
