--- driversPt2/colors.h
@@ -1,11 +1,0 @@
-#ifndef COLORS_H
-#define COLORS_H
-
-#define WHITE_ON_BLACK 0x0f
-#define RED_ON_WHITE 0xf4
-#define BLUE_ON_BLACK 0x09
-#define WHITE_ON_WHITE 0xff
-#define CYAN_ON_CYAN 0x33
-#define BLACK_ON_BLACK 0x00
-
-#endif
