--- driversPt2/sound.c
@@ -1,39 +1,0 @@
-#include "sound.h"
-//#include "../cpuPt2/types.h"
-#include "../cpuPt2/ports.h"
-#include "../cpuPt2/timer.h"
-
-void play(uint32_t nFrequence) {
-    uint32_t Div;
-    uint8_t tmp;
-
-    Div = 1193180 / nFrequence;
-    port_byte_out(0x43, 0xb6);
-    port_byte_out(0x42, (uint8_t) (Div) );
-    port_byte_out(0x42, (uint8_t) (Div >> 8));
-
-    tmp = port_byte_in(0x61);
-    if (tmp != (tmp | 3)) {
-        port_byte_out(0x61, tmp | 3);
-    }
-}
-
-void nosound() {
-    uint8_t tmp = port_byte_in(0x61) & 0xFC;
-
-    port_byte_out(0x61, tmp);
-}
-
-void play_sound(uint32_t nFrequence, uint32_t ticks) {
-    play(nFrequence);
-    wait(ticks);
-    nosound();
-}
-
-void pgw() {
-    play_sound(300, 30);
-    play_sound(400, 30);
-    wait(10);
-    play_sound(500, 30);
-    play_sound(600, 30);
-}
