+++ kernel/kernel.c
@@ -1,9 +1,13 @@
+//0x1000
 #include "../cpu/isr.h"
 #include "../drivers/screen.h"
+#include "../drivers/sound.h"
 #include "kernel.h"
 #include "../libc/string.h"
 #include "../libc/mem.h"
 #include "../cpu/timer.h"
+#include "terminal.h"
+#include "../drivers/stdin.h"
 #include "../libc/stdio.h"
 #include "../fs/hdd.h"
 #include "../fs/hddw.h"
@@ -21,45 +25,68 @@ void main() {
 	init_timer(1);
 	clear_screen();
 	empty_sector();
+	//ata_pio28(ata_controler, 1, ata_drive, 0x1);
 	prevtick = tick;
-	kprint("Loading DripOS...\n"); //Version
-	driveLoop();
+	logoDraw();
+	wait(100);
+	clear_screen();
+	kprint("DripOS 0.0012\n"); //Version
+	check_crash();
+    kprint("Type help for commands\nType shutdown to shutdown\n> ");
+	stdin_init();
+	backspace(key_buffer);
+	play_sound(500, 100);
+	play_sound(300, 100);
 }
 
-void driveLoop() {
-	uint16_t sectors = 31;
-	uint16_t sector = 0;
-	uint16_t basesector = 40000;
-	uint32_t i = 40031;
-	uint16_t code[sectors][256];
-	int x = 0;
-	while(x==0) {
-		read(i);
-		for (int p=0; p < 256; p++) {
-			if (readOut[p] == 0) {
-			} else {
-				x = 1;
-				//kprint_int(i);
-			}
-		}
-		i++;
+void user_input(char *input) {
+	if (stdinpass == 0){
+		execute_command(input);
 	}
-	kprint("Found sector!\n");
-	kprint("Loading OS into memory...\n");
-	for (sector=0; sector<sectors; sector++) {
-		read(basesector+sector);
-		for (int p=0; p<256; p++) {
-			code[sector][p] = readOut[p];
-		}
+	else {
+		stdinpass = 0;
+		stdin_call(input);
 	}
-	kprint("Done loading.\n");
-	__builtin___clear_cache(&code[0][0], &code[sectors][256]);   // don't optimize away stores into the buffer
-    void (*fptr)(void) =  (void*)code;                     // casting to void* instead of the actual target type is simpler
-	kprint("The OS is executing now.\n If nothing prints then the execution hanged.\n");
-    fptr();
-	kprint("Execution did nothing!\n");
+}
+
+void halt() {
+	asm volatile("hlt");
+}
+
+void shutdown() {
+	kprint("System shutdown");
+	state = 1;
+}
+
+void panic() {
+	state = 2;
 }
 
 int getstate() {
 	return state;
+}
+
+void memory() {
+	/* Lesson 22: Code to test kmalloc, the rest is unchanged */
+        uint32_t phys_addr;
+        uint32_t page = kmalloc(1000, 1, &phys_addr);
+        char page_str[16] = "";
+        hex_to_ascii(page, page_str);
+        char phys_str[16] = "";
+        hex_to_ascii(phys_addr, phys_str);
+        kprint("Page: ");
+        kprint(page_str);
+        kprint(", physical address: ");
+        kprint(phys_str);
+        kprint("\n");
+}
+
+void check_crash() {
+	//0x7263
+	read(128);
+	if (readOut[0] == 0x7263) {
+		kprint("NOTICE: Last time your OS stopped, it was from a crash.\n");
+	}
+	writeIn[0] = 0x0000;
+	writeFromBuffer(128);
 }
