+++ kernel/terminal.h
