--- libcPt2/function.h
@@ -1,8 +1,0 @@
-#ifndef FUNCTION_H
-#define FUNCTION_H
-
-/* Sometimes we want to keep parameters to a function for later use
- * and this is a solution to avoid the 'unused parameter' compiler warning */
-#define UNUSED(x) (void)(x)
-
-#endif
