--- libcPt2/stdio.h
@@ -1,4 +1,0 @@
-#pragma once
-void printf(char *text);
-void stdin_call(char *text);
-char * scanf(char *message);
